requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
black==22.10.0
Brotli==1.0.9
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
ConfigArgParse==1.5.3
Django==4.1.1
django-bootstrap-v5==1.0.11
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-environ==0.9.0
exceptiongroup==1.0.4
flake8==5.0.4
Flask==2.2.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==22.10.2
geventhttpclient==2.0.8
greenlet==2.0.1
hiredis==2.0.0
idna==3.4
importlib-metadata==5.1.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
locust==2.13.1
MarkupSafe==2.1.1
mccabe==0.7.0
msgpack==1.0.4
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.2
platformdirs==2.5.4
pluggy==1.0.0
psutil==5.9.4
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytest==7.2.0
pytz==2022.2.1
pyzmq==24.0.1
redis==4.3.5
requests==2.28.1
roundrobin==0.0.4
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.11.0
zope.event==4.5.0
zope.interface==5.5.2
